[not verified] Protect: Move NavigationItem to own file
